server/api/requirements.txt
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyMySQL==1.0.3
python-dotenv==1.0.0
Werkzeug==2.2.3
